+++ gui/src/helpers/database/__init__.py
@@ -2,6 +2,7 @@ from .embedding_worker import ImageEmbeddingWorker
 from .library_session import get_library_db, run_migration_with_progress
 from .recommendation_worker import RecommendationWorker
 from .search_worker import SearchWorker
+from .semantic_search_worker import SemanticSearchWorker
 from .tag_completer import TagCompleter
 from .upsert_worker import UpsertWorker
 
@@ -11,6 +12,7 @@ __all__ = [
     "ImageEmbeddingWorker",
     "RecommendationWorker",
     "SearchWorker",
+    "SemanticSearchWorker",
     "TagCompleter",
     "UpsertWorker",
 ]